Use GoReleaser to build Docker containers
This commit swaps to using GoReleaser to build & publish releases for pkdns and pkdns-cli for macOS (x64 & ARM), Linux (x64 & ARM) and Windows x64.

GoReleaser also produces Docker containers and manifests that can be used as explained in the README.

The `Dockerfile` in the repo had to be slightly altered to support this.